+++ docs/fr_FR/Changelog.rst
@@ -7,6 +7,7 @@ ChangeLog
   - Correction de certains mauvais type d'attribut pour configureReporting2.
   - Améliorations outil de check.
 - Page santé: Réactivation raffraichissement toutes les 2 sec.
+- Interne: Configuration des Zigates au démarrage revue. Déplacé dans AbeilleCmd.
 
 231205-BETA-6
 -------------